Treesame-Commit-Id: 2351dac5f42c5acb7c7645f1383dbc91750b3172

+++ app/controllers/profile_controller.rb
@@ -423,13 +423,13 @@ class ProfileController < ApplicationController
 
     short_name = params[:user] && params[:user][:short_name]
     @user.short_name = short_name if short_name && @user.user_can_edit_name?
-    if params[:user_profile]
+    if params[:user_profile] && @user.user_can_edit_profile?
       user_profile_params = params[:user_profile].permit(:title, :bio)
       user_profile_params.delete(:title) unless @user.user_can_edit_name?
       @profile.attributes = user_profile_params
     end
 
-    if params[:link_urls] && params[:link_titles]
+    if params[:link_urls] && params[:link_titles] && @user.user_can_edit_profile?
       @profile.links = []
       params[:link_urls].zip(params[:link_titles])
                         .reject { |url, title| url.blank? && title.blank? }
@@ -492,7 +492,7 @@ class ProfileController < ApplicationController
   end
 
   def content_shares
-    raise not_found unless @current_user.can_content_share?
+    raise not_found unless @current_user.can_view_content_shares?
 
     @user ||= @current_user
     set_active_tab "content_shares"
